move pathStringUtils to domain
